#fix memory leak

+++ include/buffer.h
@@ -2,11 +2,11 @@
 #include <stdint.h>
 #include <stdio.h>
 typedef void *Buffer;
-typedef Buffer (*Buffer_Alloc)(size_t size);
+typedef Buffer (*Buffer_Alloc)(uint32_t size);
 typedef void (*Buffer_Free)(Buffer buffer);
 void Buffer_init(Buffer_Alloc alloc, Buffer_Free free);
-Buffer Buffer_alloc(size_t size);
-Buffer Buffer_from(size_t size, Buffer source);
+Buffer Buffer_alloc(uint32_t size);
+Buffer Buffer_from(uint32_t size, Buffer source);
 void Buffer_free(Buffer buf);
 
 #define toBuffer(t) Buffer_from(sizeof(t), &(t))
