+++ src/util/buffer.c
@@ -2,7 +2,6 @@
 #include <stdlib.h>
 #include <string.h>
 
-
 static Buffer_Alloc g_alloc = malloc;
 static Buffer_Free g_free = free;
 
@@ -10,8 +9,8 @@ void Buffer_init(Buffer_Alloc alloc, Buffer_Free free) {
   g_alloc = alloc;
   g_free = free;
 }
-Buffer Buffer_alloc(size_t size) { return g_alloc(size); }
-Buffer Buffer_from(size_t size, Buffer source) {
+Buffer Buffer_alloc(uint32_t size) { return g_alloc(size); }
+Buffer Buffer_from(uint32_t size, Buffer source) {
   void *buf = g_alloc(size);
   memcpy(buf, source, size);
   return buf;
